+++ src/App.tsx
@@ -2,7 +2,7 @@ import Router from './components/Router';
 import styled, { ThemeProvider } from 'styled-components';
 import GlobalStyle from './styles/GlobalStyle';
 import theme from './styles/theme';
-import { useEffect } from 'react';
+import { useSetScreenSizeOnResizeEvent } from './hooks/useSetScreenSizeOnResizeEvent';
 
 const Wrapper = styled.div`
   position: relative;
@@ -16,27 +16,7 @@ const Wrapper = styled.div`
 `;
 
 function App() {
-  function setScreenSize() {
-    //vh 화면 비율 적용
-    const vh = window.innerHeight * 0.01;
-    document.documentElement.style.setProperty('--vh', `${vh}px`);
-
-    // vw 화면 비율 적용
-    const windowWidth =
-      window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
-    const maxWidth = Math.min(375, windowWidth);
-    document.documentElement.style.setProperty('--app-max-width', `${maxWidth}px`);
-  }
-
-  useEffect(() => {
-    setScreenSize();
-    window.addEventListener('resize', setScreenSize);
-
-    return () => {
-      window.removeEventListener('resize', setScreenSize);
-    };
-  }, []);
-
+  useSetScreenSizeOnResizeEvent();
   return (
     <Wrapper>
       <ThemeProvider theme={theme}>
