+++ src/hooks/useSetScreenSizeOnResizeEvent.ts
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+
+export function useSetScreenSizeOnResizeEvent() {
+  function setScreenSize() {
+    //vh 화면 비율 적용
+    const vh = window.innerHeight * 0.01;
+    document.documentElement.style.setProperty('--vh', `${vh}px`);
+
+    // vw 화면 비율 적용
+    const windowWidth =
+      window.innerWidth || document.documentElement.clientWidth || document.body.clientWidth;
+    const maxWidth = Math.min(375, windowWidth);
+    document.documentElement.style.setProperty('--app-max-width', `${maxWidth}px`);
+  }
+
+  useEffect(() => {
+    setScreenSize();
+    window.addEventListener('resize', setScreenSize);
+
+    return () => {
+      window.removeEventListener('resize', setScreenSize);
+    };
+  }, []);
+}
